requirements.txt
annotated-types==0.5.0
anyio==3.7.0
Authlib==1.2.1
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
ecdsa==0.18.0
exceptiongroup==1.1.2
fastapi==0.99.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
passlib==1.7.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.11
pydantic_core==2.0.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
